% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-logger.R
\name{.initLogger}
\alias{.initLogger}
\title{Sets up the lgr logger hierarchy with JSON file output.}
\usage{
.initLogger(log_dir = "logs/files")
}
\arguments{
\item{log_dir}{directory for log files}
}
\description{
Sets up the lgr logger hierarchy with JSON file output.
}
\keyword{internal}
