% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-logger.R
\name{.logWarning}
\alias{.logWarning}
\title{Log warning both in lgr and base R warning}
\usage{
.logWarning(msg, ...)
}
\arguments{
\item{msg}{message to be printed}

\item{...}{further arguments for lgr}
}
\description{
Log warning both in lgr and base R warning
}
\keyword{internal}
