% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-db.R
\name{.sqlDropTable}
\alias{.sqlDropTable}
\title{Drop a database table if it exists}
\usage{
.sqlDropTable(con, table_name)
}
\arguments{
\item{con}{A DBI connection.}

\item{table_name}{Character. Name of the table to drop.}

\item{lg}{An lgr logger object.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Drop a database table if it exists
}
\keyword{internal}
