% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-db.R
\name{.timeToDays}
\alias{.timeToDays}
\title{Convert a timepoint to days}
\usage{
.timeToDays(days = 0, weeks = 0, months = 0, years = 0)
}
\arguments{
\item{days}{Number of days.}

\item{weeks}{Number of weeks.}

\item{months}{Number of months.}

\item{years}{Number of years.}
}
\value{
A numeric value representing the total number of days.
}
\description{
Converts a timepoint expressed in days, weeks, months, or years to a total
number of days. Multiple units are summed. Assumes 1 year = 365 days,
1 month = 30 days, 1 week = 7 days.
}
\keyword{internal}
