% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{FAIRgroundDisconnect}
\alias{FAIRgroundDisconnect}
\title{Disconnect the default FAIRground database connection}
\usage{
FAIRgroundDisconnect()
}
\value{
Called for its side effects. Invisibly returns \code{TRUE}.
}
\description{
Closes the package default connection stored by \code{\link[=FAIRgroundConnect]{FAIRgroundConnect()}} and
clears it from the internal package environment.
}
\examples{
\dontrun{
FAIRgroundConnect("LUCID_FAIRPOINT")
FAIRgroundDisconnect()
}
}
\seealso{
\code{\link[=FAIRgroundConnect]{FAIRgroundConnect()}}, \code{\link[=FAIRgroundConnection]{FAIRgroundConnection()}}
}
