% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table_list.R
\name{filterTableList}
\alias{filterTableList}
\title{Filter a table list by study IDs}
\usage{
filterTableList(tableList, study_ids)
}
\arguments{
\item{tableList}{Named list of data frames as produced by \code{\link[=getDatabase]{getDatabase()}}.}

\item{study_ids}{Character vector of study identifiers to keep (e.g. \code{c("CHHIL", "ITCHI")}).}
}
\value{
A filtered version of \code{tableList}, containing only rows linked to the selected studies.
}
\description{
This function takes an existing tableList object (as returned by getDatabase())
and filters all tables based on one or more study_id's. It preserves the structure of
the dataset while restricting its contents to only relevant studies, experiments and analyses.
}
\examples{
\dontrun{
  data <- getDatabase(path = "data/")
  filtered <- filterTableList(data, study_ids = c("CHHIL", "ITCHI"))
}

}
\seealso{
\code{\link[=getDatabase]{getDatabase()}}, \code{\link[=databaseExport]{databaseExport()}}
}
