% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_gates.R
\name{harmonizeGates}
\alias{harmonizeGates}
\title{Takes a vector of gate strings, splits them into components, harmonizes the marker names,
and reconstructs the gate strings with standardized marker names using \code{harmonizeMarkers}.}
\usage{
harmonizeGates(gates)
}
\arguments{
\item{gates}{A character vector where each element is a gate string (e.g., "HLA-DR_+\emph{OR_CD7}-").}
}
\value{
A character vector of harmonized gate strings. Each output corresponds to one input gate.
}
\description{
Takes a vector of gate strings, splits them into components, harmonizes the marker names,
and reconstructs the gate strings with standardized marker names using \code{harmonizeMarkers}.
}
\examples{
harmonizeGates(c("HLA-DR_+_OR_CD7_-", "CD3_+", "CD16_-_AND_CD56_++"))
}
