% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{startProject}
\alias{startProject}
\title{Create a FAIRground folder}
\usage{
startProject(dir, dummy = FALSE)
}
\arguments{
\item{dir}{Character. Path where the FAIRground should be
created. The directory will be created if it does not already exist.}

\item{dummy}{Logical. If \code{TRUE}, the project is populated with example
dummy data and scripts included in the package. Useful for vignettes,
demonstrations, or learning the FAIRground. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effects on the file system.
}
\description{
Initializes a FAIRground directory containing the standard folder
structure required for ETL development and database integration. The
following top-level directories are created:
\itemize{
\item \verb{studies/}
\item \verb{experiments/}
\item \verb{analyses/}
\item \verb{common/}
\item \verb{logs/}
\item \verb{codebooks/custom/}
}
}
\details{
These directories are intended to contain ETL scripts (typically located
in \verb{scripts/} subfolders) that extract raw data and load them into the
connected FAIRground database.

In addition, the function:
\itemize{
\item Creates an RStudio project file (\code{.Rproj})
\item Generates a basic \code{.gitignore} file
}

Optionally, the project can be populated with a bundled dummy example
project to demonstrate the FAIRground.

Creates the  FAIRground directory structure on disk and optionally
copies example files bundled with the package. To start drafting ETL scripts
and populating the FAIRground use \code{\link[=addStudy]{addStudy()}}, \code{\link[=addExperiment]{addExperiment()}},
and \code{\link[=addAnalysis]{addAnalysis()}}

The function does not initialize a database connection. After creating the
project structure, use \code{\link[=FAIRgroundConnect]{FAIRgroundConnect()}} to connect to a database.
}
\examples{
\dontrun{
startProject(dir = "C:/Users/johndoe/Projects/LUCID_beFAIR")

# Setup a dummy project
dir <- paste0(tempdir(), "/dummy_beFAIR")
startProject(dir = dir, dummy=TRUE)

# Inspect the created beFAIR directory
cat(normalizePath(dir), "\n")
}

}
\seealso{
\code{\link[=FAIRgroundConnect]{FAIRgroundConnect()}}, \code{\link[=addStudy]{addStudy()}}, \code{\link[=addExperiment]{addExperiment()}},
\code{\link[=addAnalysis]{addAnalysis()}}
}
